requirements.txt
antlr4-python3-runtime
hydra-core
omegaconf
packaging
PyYAML
albucore
albumentations
annotated-types
antlr4-python3-runtime
asttokens
attrs
backcall
beautifulsoup4
bleach
certifi
charset-normalizer
click
comm
contourpy
cycler
debugpy
decorator
defusedxml
docker-pycreds
docopt
einops
einx
exceptiongroup
executing
fastjsonschema
filelock
fonttools
frozendict
fsspec
future
gitdb
GitPython
hydra-core
idna
imageio
ipykernel
ipython
jedi
Jinja2
joblib
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab_pygments
kiwisolver
lazy_loader
MarkupSafe
matplotlib
matplotlib-inline
mistune
monai
mpmath
natsort
nbclient
nbconvert
nbformat
ncps
nest-asyncio
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
omegaconf
opencv-python
opencv-python-headless
packaging
pandas
pandocfilters
parso
pexpect
pickleshare
pillow
platformdirs
prompt_toolkit
protobuf
psutil
ptyprocess
pure_eval
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
pytz
PyYAML
pyzmq
referencing
requests
rpds-py
safetensors
scikit-image
scikit-learn
scipy
seaborn
sentry-sdk
setproctitle
simsimd
six
smmap
soupsieve
stack-data
stringzilla
sympy
threadpoolctl
tifffile
tinycss2
torch
torchvision
tornado
tqdm
traitlets
triton
typing_extensions
tzdata
urllib3
vector-quantize-pytorch
wandb
wcwidth
webencodings
yarg
